v.1

+++ lib/services/firebase_service.dart
@@ -6,6 +6,7 @@ class FirebaseService {
   final FirebaseAuth _auth = FirebaseAuth.instance;
   final FirebaseFirestore _firestore = FirebaseFirestore.instance;
 
+  User? get currentUser => _auth.currentUser;
   // Autenticación
   Future<UserCredential> signIn(String email, String password) async {
     try {
@@ -29,9 +30,22 @@ class FirebaseService {
     }
   }
 
+  // Cierre de sesión
+  Future<void> signOut() async {
+    try {
+      await _auth.signOut();
+    } catch (e) {
+      throw Exception('Error al cerrar sesión: $e');
+    }
+  }
+
   // CRUD Operaciones
   Future<void> addVehicle(ElectricVehicle vehicle) async {
+    if (currentUser == null) throw Exception('Usuario no autenticado');
+
     try {
+      // Aseguramos que el userId sea el del usuario actual
+      vehicle.userId = currentUser!.uid;
       await _firestore.collection('vehicles').add(vehicle.toMap());
     } catch (e) {
       throw Exception('Error al agregar vehículo: $e');
@@ -39,15 +53,30 @@ class FirebaseService {
   }
 
   Stream<List<ElectricVehicle>> getVehicles() {
-    return _firestore.collection('vehicles').snapshots().map((snapshot) {
-      return snapshot.docs.map((doc) {
-        return ElectricVehicle.fromMap(doc.data(), doc.id);
-      }).toList();
+    if (currentUser == null) return Stream.value([]);
+
+    return _firestore
+        .collection('vehicles')
+        .where('userId', isEqualTo: currentUser!.uid)
+        .snapshots()
+        .map((snapshot) {
+      return snapshot.docs
+          .map((doc) => ElectricVehicle.fromMap(doc.data(), doc.id))
+          .toList();
     });
   }
 
   Future<void> updateVehicle(ElectricVehicle vehicle) async {
+    if (currentUser == null) throw Exception('Usuario no autenticado');
+
     try {
+      // Verificar que el vehículo pertenece al usuario actual
+      final vehicleDoc =
+          await _firestore.collection('vehicles').doc(vehicle.id).get();
+      if (vehicleDoc.data()?['userId'] != currentUser!.uid) {
+        throw Exception('No tienes permiso para modificar este vehículo');
+      }
+
       await _firestore
           .collection('vehicles')
           .doc(vehicle.id)
@@ -58,7 +87,16 @@ class FirebaseService {
   }
 
   Future<void> deleteVehicle(String vehicleId) async {
+    if (currentUser == null) throw Exception('Usuario no autenticado');
+
     try {
+      // Verificar que el vehículo pertenece al usuario actual
+      final vehicleDoc =
+          await _firestore.collection('vehicles').doc(vehicleId).get();
+      if (vehicleDoc.data()?['userId'] != currentUser!.uid) {
+        throw Exception('No tienes permiso para eliminar este vehículo');
+      }
+
       await _firestore.collection('vehicles').doc(vehicleId).delete();
     } catch (e) {
       throw Exception('Error al eliminar vehículo: $e');
